requirements.txt
slackclient==2.8.0
click==7.1.2
